requirements.txt
# FastAPI and web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.0

# FastMCP for MCP servers
fastmcp==0.1.0

# OpenAI Agents SDK
openai-agents==0.0.3

# Anthropic - Claude API
anthropic==0.45.0

# OpenAI - for Agents SDK and fallbacks
openai>=1.66.2

# Database
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.35
alembic==1.14.0

# Redis
redis==5.0.8
aioredis==2.0.1

# Telephony
twilio==9.3.0

# Speech-to-Text
deepgram-sdk==3.6.0

# Text-to-Speech
elevenlabs==1.5.0

# WebSockets
websockets==13.1

# Utilities
python-dotenv==1.0.1
pydantic>=2.10
pydantic-settings>=2.6.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
